update tests for publish
